# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Prodman API
  version: '1'
  description: This is the API documentation for Prodman, a production management tool.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local server
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
...
paths:
...
  /auth/session:
    get:
      operationId: session
      tags:
        - Auth
      summary: Endpoint to check if session is active
      responses:
        '200':
          description: Successful operation
        '440':
          description: Login timeout
        '500':
          description: Internal server error
...